@@ -1,16 +1,16 @@
-import fs = require("fs");
+import * as fs from "fs";
 import { IHoldings, ITrade } from "../types";
 
 export async function save(holdings: IHoldings, trades: ITrade[]): Promise<any> {
     const data: string = JSON.stringify({
-        savedDate: new Date,
+        savedDate: new Date(),
         holdings,
         trades,
     });
-    return new Promise(function(resolve: Function, reject: Function): void {
-        fs.writeFile("./data.json", data, function(err: Error): void {
+    return new Promise((resolve: () => void, reject: (Error) => void): void => {
+        fs.writeFile("./data.json", data, (err: Error): void => {
             if (err) {
-                console.log(err);
+                // console.log(err);
                 reject(err);
             }
             resolve();
